Task Coach would sometimes crash when dragging a task onto another task on Windows. Fixed by using the wx.lib.agw.hypertreelist.HyperTreeList widget for tree-list controls. Since a number of bugs have been fixed in HyperTreeList recently (some of which found and fixed due to us using it in Task Coach) recent copies of HyperTreeList and CustomTreeCtrl have been included in taskcoachlib/thirdparty.
BTW, there is still a bug in HyperTreeList.HitTest that I hope will be fixed by Andrea Gavana (HyperTreeList author) in the near future.

Since we are now using the HyperTreeList widget, it was easy to extend the category viewer to show additional columns (description, attachment, ...) besides just the subject of categories.