More notification-related refactorings. Most methods that change an attribute of a domain object or container now look like this:
def setLocation(self, location, event=None):
    if location == self.__location:
        return event
    notify = event is None
    event = event or patterns.Event()
    self.__location = location
    event.addSource(self, location, type=self.locationChangedEventType())
    if notify:
        event.send()
    else:
        return event

If the method is called *without* an event, it creates one itself and notifies observers at the end by sending the event. If the method is called *with* an event it will add its changes to the event and return the changed event. Returning is not strictly necessary as the method is changing a mutable object (the event instance) but this makes it more clear in the resulting code that the event is being changed. All this allows for doing multiple changes in a row with just one event being sent. For example:

event = patterns.Event()
attachment.setLocation(newLocation, event)
task.setSubject(newSubject, event)
event.send()

There is now a massive amount of duplication in the source code since all "setters" now look like the method above. Next step will be to remove that duplication, either by introducing an attribute class that takes care of notifications, or a decorator that does the work.