Handle long integers properly for all Python versions
In Python 2, `int` is a native word-sized integer up to 2³¹−1 on 32-bit
platforms and 2⁶³−1 on 64-bit platforms, while `long` is an arbitrary
precision integer. In Python 3, `int` is an arbitrary precision integer and
`long` does not exist. Initialize the `ordering` attribute to whatever is the
appropriate arbitrary precision integer type on the current Python
implementation.